% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{available_locations}
\alias{available_locations}
\title{Get list of localities supported by Zomato}
\usage{
available_locations()
}
\value{
A list of locations with latitudes and longitudes
}
\description{
Get list of localities supported by Zomato
}
