% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_weather}
\alias{get_weather}
\title{Get current weather at a particular latitude and longitude}
\usage{
get_weather(api_key, latitude, longitude)
}
\arguments{
\item{api_key}{API key}

\item{latitude}{Latitude}

\item{longitude}{Longitude}
}
\value{
A dataframe with current weather statistics
including: temperature, humidity, wind_speed, wind_direction
rain_intensity rain_accumulation latitude
}
\description{
Get current weather at a particular latitude and longitude
}
