% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to `zomato_locations.csv` file}
\source{
This function is adapted from `readxl::readxl_example()`.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
`"zomato_locations.csv"` will work.
If `NULL`, the example files will be listed.}
}
\description{
weatherunion comes bundled with a csv file in `inst/extdata`
directory. This function make them easy to access.
}
\examples{
\dontrun{
path_to_file()
path_to_file("zomato_locations.csv")
head(read.csv(path_to_file("zomato_locations.csv")))
}
}
