% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zomato_locations}
\alias{zomato_locations}
\title{Zomato locations data}
\format{
## `zomato_locations`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{cityName}{City}
  \item{localityName}{locality}
  \item{localityId}{Location ID}
  \item{latitude}{Latitude}
  \item{longitude}{Longitude}
  \item{device_type}{Device type}
  ...
}
}
\source{
<https://b.zmtcdn.com/data/file_assets/65fa362da3aa560a92f0b8aeec0dfda31713163042.pdf>
}
\usage{
zomato_locations
}
\description{
List of localities supported by Zomato
Report ...
}
\keyword{datasets}
